Adopt some clang-tidy review suggestions #3368

+++ cpp/core/progressbar.h
@@ -243,8 +243,8 @@ FORCE_INLINE void ProgressBar::operator++() {
     }
   } else {
     if (++current_val >= next_percent) {
-      _value = static_cast<size_t>(std::round(current_val / _multiplier));
-      next_percent = static_cast<size_t>(std::ceil((_value + 1) * _multiplier));
+      _value = static_cast<size_t>(std::round(static_cast<float>(current_val) / _multiplier));
+      next_percent = static_cast<size_t>(std::ceil(static_cast<float>(_value + 1) * _multiplier));
       display_now();
     }
   }
